+++ CHANGELOG.md
@@ -32,7 +32,7 @@ All notable changes to this project will be documented in this file.<br>
 
 ## Master
 ### Added
-- Nothing
+- Added `setTabBarVisible(_:, animated:)` and `isTabBarVisible` in UIViewController extension
 
 ### Improved
 - Updated SwiftLint to 0.21.0